chore : 좌석 더미 데이터 추가
